requirements.txt
# Core dependencies
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic-ai>=0.0.13
typing-extensions>=4.7.0

# GPU Hardware Integration
pynvml>=11.5.0  # NVIDIA GPU management
# rocm-smi-lib>=5.6.0  # AMD GPU management (system package)
# py-level-zero>=1.0.0  # Intel GPU management (when available)

# Distributed Systems
grpcio>=1.54.0
grpcio-tools>=1.54.0
protobuf>=4.23.0
aiohttp>=3.8.0
websockets>=11.0.0

# Real-time and Performance
asyncio>=3.4.3
uvloop>=0.17.0  # High-performance event loop
cython>=3.0.0  # For performance-critical components
numba>=0.57.0  # JIT compilation for numerical code

# Monitoring and Observability
prometheus-client>=0.17.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
structlog>=23.1.0
rich>=13.4.0  # Rich terminal output

# Security and Authentication
cryptography>=41.0.0
jwt>=1.3.1
bcrypt>=4.0.1

# API and Networking
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
starlette>=0.27.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-benchmark>=4.0.0
pytest-xdist>=3.3.0  # Parallel test execution

# Code Quality
black>=23.7.0
mypy>=1.5.0
ruff>=0.0.280
pre-commit>=3.3.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0
mkdocs-mermaid2-plugin>=1.1.0